Add files via upload
Changed the structure of the code so that playLog and playEvent are now seperate classes capable of storing select properties. PlayLog will handle the majority of the processing while PlayEvent handles the important event info such as startTime and endTime. The list Content holds all the necessary data and is accessed through the PlayLog class. The main function only acts as an XmlReader initialiser and as a potential interface for the user to control.